+++ commands/compose/attach.go
@@ -1,11 +1,14 @@
 package compose
 
 import (
+	"errors"
 	"fmt"
 	"os"
+	"path/filepath"
 	"strings"
 
 	"git.sr.ht/~rjarry/aerc/commands"
+	"git.sr.ht/~rjarry/aerc/logging"
 	"git.sr.ht/~rjarry/aerc/widgets"
 	"github.com/mitchellh/go-homedir"
 )
@@ -31,26 +34,45 @@ func (Attach) Execute(aerc *widgets.Aerc, args []string) error {
 	}
 
 	path := strings.Join(args[1:], " ")
-
 	path, err := homedir.Expand(path)
 	if err != nil {
+		logging.Errorf("failed to expand path '%s': %v", path, err)
 		aerc.PushError(err.Error())
 		return err
 	}
 
-	pathinfo, err := os.Stat(path)
-	if err != nil {
-		aerc.PushError(err.Error())
-		return err
-	} else if pathinfo.IsDir() {
-		aerc.PushError("Attachment must be a file, not a directory")
-		return nil
+	logging.Debugf("attaching %s", path)
+
+	attachments, err := filepath.Glob(path)
+	if err != nil && errors.Is(err, filepath.ErrBadPattern) {
+		logging.Warnf("failed to parse as globbing pattern: %v", err)
+		attachments = []string{path}
 	}
+
+	logging.Debugf("filenames: %v", attachments)
 
 	composer, _ := aerc.SelectedTabContent().(*widgets.Composer)
-	composer.AddAttachment(path)
+	for _, attach := range attachments {
+		logging.Debugf("attaching '%s'", attach)
 
-	aerc.PushSuccess(fmt.Sprintf("Attached %s", pathinfo.Name()))
+		pathinfo, err := os.Stat(attach)
+		if err != nil {
+			logging.Errorf("failed to stat file: %v", err)
+			aerc.PushError(err.Error())
+			return err
+		} else if pathinfo.IsDir() && len(attachments) == 1 {
+			aerc.PushError("Attachment must be a file, not a directory")
+			return nil
+		}
+
+		composer.AddAttachment(attach)
+	}
+
+	if len(attachments) == 1 {
+		aerc.PushSuccess(fmt.Sprintf("Attached %s", path))
+	} else {
+		aerc.PushSuccess(fmt.Sprintf("Attached %d files", len(attachments)))
+	}
 
 	return nil
 }